services/aof-agent/requirements.txt
websockets==11.0.3
nats-py==2.6.0
asyncio==3.4.3